fix: ファイル削除をzipファイル参照からディレクトリ参照に変更

+++ src/extractor.rs
@@ -1,7 +1,7 @@
 use std::error::Error;
-use std::fs::{create_dir_all, remove_dir_all, remove_file, File};
+use std::fs::{create_dir_all, read_dir, remove_dir_all, remove_file, File};
 use std::io::{Cursor, copy, Read, Seek};
-use std::path::{Path, PathBuf};
+use std::path::{PathBuf};
 use bytes::Bytes;
 use zip::ZipArchive;
 
@@ -21,23 +21,19 @@ impl Extractor {
     pub fn extract(&self, data :Bytes) -> Result<(), Box<dyn Error>> {
         let cursor = Cursor::new(data.as_ref());
         let archive = ZipArchive::new(cursor)?;
-        let mut remove_name = archive.file_names().filter(|&n| !self.ignore.contains(&n.to_string())).collect::<Vec<&str>>();
-        remove_name.dedup_by(|a, b| {
-            let a_v = Path::new(a).ancestors().collect::<Vec<_>>();
-            let b_v = Path::new(b).ancestors().collect::<Vec<_>>();
-            a_v.get(a_v.len() -2) == b_v.get(b_v.len() -2)
-        });
-
-        self.remove_all(remove_name)?;
+        self.remove_all()?;
         self._extract(archive)?;
 
         Ok(())
     }
 
-    fn remove_all(&self, names :Vec<&str>) -> Result<(), Box<dyn Error>> {
-        for name in names {
-            let path_str = format!("{}/{}", self.output_dir, name);
-            let path = Path::new(path_str.as_str());
+    pub fn remove_all(&self) -> Result<(), Box<dyn Error>> {
+        let paths = read_dir(&self.output_dir)?;
+        for path in paths {
+            let path = path?.path();
+            if path.file_name().is_some_and(|n| self.ignore.contains(&n.to_string_lossy().to_string())) {
+                continue
+            }
             if path.exists() {
                 if path.is_dir() {
                     remove_dir_all(&path)?;
